refactored Events table. Deleted Get_events table
