+++ db/migrate/20200210180900_create_events.rb
@@ -2,10 +2,10 @@ class CreateEvents < ActiveRecord::Migration[5.2]
   def change
     create_table :events do |t|
       t.string :title
+      t.string :date
       t.string :category
-      t.date :date
-      t.string :venue
-      t.string :ticket_info
+      t.string :event_id_num
+      t.string :description
     end
   end
 end
