--- db/migrate/20200210180900_create_get_events.rb
@@ -1,10 +1,0 @@
-class CreateGetEvents < ActiveRecord::Migration[5.2]
-  def change
-    create_table :get_events do |t|
-      t.string :title
-      t.string :category
-      t.string :event_id_num
-      t.string :description
-    end
-  end
-end
