+++ db/seeds.rb
@@ -1,7 +1,3 @@
 User.create(first_name: "Bob", last_name: "Smith", age: 27, gender:"m")
 User.create(first_name: "Ann", last_name: "Jones", age: 34, gender:"f")
 User.create(first_name: "Pete", last_name: "Hale", age: 23, gender:"m")
-
-Event.create(title: "Ultimate Frisbee", category: "outdoors", date: 2020-02-14, venue: "Zilker Park", ticket_info: "free")
-Event.create(title: "Gary P Clarke", category: "music", date: 2020-02-14, venue: "Emo's", ticket_info: "23")
-Event.create(title: "Jimmy Vaughn", category: "music", date: 2020-02-14, venue: "Antone's", ticket_info: "30")
