+++ lib/schedule.rb
@@ -1,0 +1,4 @@
+class Schedule < ActiveRecord::Base
+  belongs_to :user
+  belongs_to :event
+end
