+++ lib/user.rb
@@ -1,3 +1,4 @@
 class User < ActiveRecord::Base
-
+  has_many :schedules
+  has_many :events, through: :schedules
 end
